+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 1.0a11 (unreleased)
 -------------------
 
-- Fix schedule display in Contact section (days delta)
+- Fix schedule display in Contact section (days delta, format & translations)
   [laulaz]
 
 - Put subsite logo & navigation viewlets in a new viewlet manager (to have custom